Inquire Update v1.1.0 Merging Dev onto Prod

+++ client/src/components/posts/wrappers/PostWrapper.js
@@ -1,4 +1,5 @@
 import React, { useState, useContext } from "react";
+import Moment from "react-moment";
 import styled, { css } from "styled-components";
 import { UserContext } from "../../context/UserProvider";
 import { UserRoleContext } from "../../context/UserRoleProvider";
@@ -12,6 +13,7 @@ import LazyFetch from "../../common/requests/LazyFetch";
 import PinIcon from "../../../imgs/pin.svg";
 import { Link } from "react-router-dom";
 import { useWindowDimensions } from "../../common/CustomHooks";
+import MobileReactionsDropdown from "../mobile/MobileReactionsDropdown";
 
 const accentColor = (type) => {
   switch (type) {
@@ -78,20 +80,24 @@ const PostWrapper = ({
   };
 
   const handleDelete = (postId, courseId) => {
-    LazyFetch({
-      type: "delete",
-      endpoint: "/courses/" + courseId + "/posts",
-      data: { _id: postId },
-      onSuccess: (data) => {
-        // console.log("Success: ", data);
-        if (postid) history.push("/course/" + courseId);
-        else window.location.reload();
-      },
-      onFailure: (err) => {
-        alert("Error deleting post.");
-        // console.log("Err: ", err);
-      },
-    });
+    let confirm = window.confirm("Are you sure you want to delete this post?");
+
+    if (confirm) {
+      LazyFetch({
+        type: "delete",
+        endpoint: "/courses/" + courseId + "/posts",
+        data: { _id: postId },
+        onSuccess: (data) => {
+          // console.log("Success: ", data);
+          if (postid) history.push("/course/" + courseId);
+          else window.location.reload();
+        },
+        onFailure: (err) => {
+          alert("Error deleting post.");
+          // console.log("Err: ", err);
+        },
+      });
+    }
   };
 
   const handleEdit = () => {
@@ -232,7 +238,7 @@ const PostWrapper = ({
         <PostTitle style={{ cursor: "pointer" }}>
           {postObject.title ? postObject.title : "Error getting post title"}
         </PostTitle>
-        <DropDownContainer>
+        <DropDownContainer id={"dropdown-container"}>
           {pinnedStatus ? (
             <img
               src={PinIcon}
@@ -245,7 +251,7 @@ const PostWrapper = ({
           ) : (
             <></>
           )}
-          {userRole && dropdownOptions && width >= 768 ? (
+          {userRole && dropdownOptions /*&& width >= 400*/ ? (
             <Dropdown options={dropdownOptions}>
               <Icon src={OptionDots} style={{ cursor: "pointer" }} />
             </Dropdown>
@@ -273,9 +279,19 @@ const PostWrapper = ({
       <HRSeperator />
       <FooterContentWrapper>
         {postObject.postedBy.anonymous ? (
-          <UserDescription isInstructor={postObject.isInstructor}>
-            Posted by {postObject.postedBy.first} {postObject.postedBy.last}
-          </UserDescription>
+          <UserInfoWrapper>
+            <UserDescription isInstructor={postObject.isInstructor}>
+              Posted by {postObject.postedBy.first} {postObject.postedBy.last}
+            </UserDescription>
+            <Moment
+              style={{
+                margin: `0.25em 0 0 0.5em`,
+                color: `#595959`,
+                fontSize: `12px`,
+              }}
+              date={postObject.createdDate}
+            />
+          </UserInfoWrapper>
         ) : (
           <Link
             style={{
@@ -290,39 +306,61 @@ const PostWrapper = ({
             }}
           >
             <UserIcon src={postObject.postedBy.picture} />
-            <UserDescription isInstructor={postObject.isInstructor}>
-              Posted by {postObject.postedBy.first} {postObject.postedBy.last}
-            </UserDescription>
+            <UserInfoWrapper>
+              <UserDescription isInstructor={postObject.isInstructor}>
+                Posted by {postObject.postedBy.first} {postObject.postedBy.last}
+              </UserDescription>
+              <Moment
+                style={{
+                  margin: `0.25em 0 0 0.5em`,
+                  color: `#595959`,
+                  fontSize: `12px`,
+                }}
+                date={postObject.createdDate}
+              />
+            </UserInfoWrapper>
           </Link>
         )}
-        {/* <UserDescription isInstructor={postObject.isInstructor}>
-          Posted by {postObject.postedBy.first} {postObject.postedBy.last}
-        </UserDescription> */}
-        <ReactionSection>
-          {userRole && userRole.participation.reactions && (
-            <Reaction
-              reactions={postObject.reactions}
-              type="post"
-              id={postObject._id}
-              postid={postObject._id}
-            />
-          )}
+        {width >= 482 ? (
+          <ReactionSection>
+            {userRole && userRole.participation.reactions && (
+              <Reaction
+                reactions={postObject.reactions}
+                type="post"
+                id={postObject._id}
+                postid={postObject._id}
+              />
+            )}
 
-          <Icon
-            alt={"Number of comments"}
-            src={CommentImg}
-            width={"22px"}
-            style={{
-              float: "left",
-              marginRight: "8px",
-              marginLeft: "20px",
-              userSelect: "none",
-            }}
-          />
-          <h5 style={{ color: "#8c8c8c", marginRight: "1em" }}>
-            {postObject.comments}
-          </h5>
-        </ReactionSection>
+            <Icon
+              alt={"Number of comments"}
+              src={CommentImg}
+              width={"22px"}
+              style={{
+                float: "left",
+                marginRight: "8px",
+                marginLeft: "20px",
+                userSelect: "none",
+              }}
+            />
+            <h5 style={{ color: "#8c8c8c", marginRight: "1em" }}>
+              {postObject.comments}
+            </h5>
+          </ReactionSection>
+        ) : (
+          <>
+            <MobileReactionsDropdown
+              postObject={postObject}
+              userRole={userRole}
+              reactionSettings={{
+                reactions: postObject.reactions,
+                type: "post",
+                id: postObject._id,
+                postid: postObject._id,
+              }}
+            />
+          </>
+        )}
       </FooterContentWrapper>
     </Wrapper>
   );
@@ -360,7 +398,7 @@ const HeaderContentWrapper = styled.div`
   transition: 150ms ease-out;
 
   @media only screen and (max-width: 768px) {
-    flex-wrap: wrap;
+    /* flex-wrap: wrap; */
     height: auto;
   }
 `;
@@ -400,6 +438,11 @@ const DropDownContainer = styled.div`
   margin-bottom: 0.5em;
 
   /* background-color: #e7e7e7; */
+
+  @media only screen and (max-width: 768px) {
+    /* margin-left: 0; */
+    justify-self: flex-end;
+  }
 `;
 
 //#endregion
@@ -438,6 +481,13 @@ const FooterContentWrapper = styled.div`
   /* border: 1px solid orange; */
 `;
 
+const UserInfoWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  /* align-items: center; */
+  justify-content: center;
+`;
+
 const UserDescription = styled.h5`
   margin-left: 0.5em;
 
@@ -453,9 +503,9 @@ const ReactionSection = styled.div`
   height: 100%;
   align-items: center;
 
-  @media only screen and (max-width: 768px) {
+  /* @media only screen and (max-width: 481px) {
     width: 0;
     display: hidden;
     overflow: hidden;
-  }
+  } */
 `;
